pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-memory-server"
version = "0.1.0"
description = "Framework-agnostic memory system for AI agents, exposed as an MCP server"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ["mcp", "memory", "ai", "agent", "llm", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp[cli]>=1.0.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "sqlite-vec>=0.1.0",
    "aiosqlite>=0.20.0",
    "httpx>=0.27.0",
    "uvicorn>=0.30.0",
]

[project.urls]
Homepage = "https://github.com/khang/agent-memory"
Repository = "https://github.com/khang/agent-memory"
Issues = "https://github.com/khang/agent-memory/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "pytest-cov>=5.0",
    "respx>=0.21.0",
]

[project.scripts]
agent-memory-server = "agent_memory.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_memory"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
